Put auth endpoints on their own router.
